+++ benchmarks/benchmark.js
@@ -203,6 +203,7 @@ async function tests() {
         { fn: convoInstance.omnid.adaptors.getDebankData, withConfig: false },
         { fn: convoInstance.omnid.adaptors.getDeepDaoData, withConfig: true },
         { fn: convoInstance.omnid.adaptors.addressToEns, withConfig: false },
+        { fn: convoInstance.omnid.adaptors.getEmblemData, withConfig: false },
         { fn: convoInstance.omnid.adaptors.getEtherscanData, withConfig: true },
         { fn: convoInstance.omnid.adaptors.getFortaData, withConfig: false },
         { fn: convoInstance.omnid.adaptors.getFoundationData, withConfig: true },